Rename and add to list
