+++ frontend/ui/data/keyboardlayouts/ar_AA_keyboard.lua
@@ -1,5 +1,5 @@
 local en_popup = require("ui/data/keyboardlayouts/keypopup/en_popup")
-local ar_popup = require("ui/data/keyboardlayouts/keypopup/ar_popup")
+local ar_popup = require("ui/data/keyboardlayouts/keypopup/ar_AA_popup")
 local com = en_popup.com -- comma (,)
 local prd = en_popup.prd -- period (.)
 local _at = en_popup._at
